Breakout a test fixture to reduce module size.
